SOTA/REBEL/requirements.txt
omegaconf==2.0.6
PyYAML==6.0.1
pytorch-lightning==1.1.7
hydra-core==1.0.6
transformers
neptune-client==0.5.1
psutil==5.8.0
datasets==1.3.0
rouge-score==0.0.4
sacrebleu==1.5.0
wandb==0.10.26
streamlit==0.82.0

#mine
scikit-learn
matplotlib
